moved tools to bin, installable scripts from bin
