+++ bin/experiment_square.py
