+++ bin/fixcoverage.py
