+++ bin/generate_hists_from_real_data.py
