+++ bin/generate_sequence.py
