+++ bin/hist_stat.py
