+++ bin/histtrim_test.py
