+++ bin/read_simulator.py
