+++ bin/remove_first_line.sh
