+++ bin/runkhmer.sh
