+++ bin/sam_to_fasta.py
