+++ bin/williams_table.py
