+++ covest/tools/experiment_parser.py
@@ -4,7 +4,7 @@ import glob
 import os
 import json
 from collections import defaultdict
-from table_generator import format_table
+from .table_generator import format_table
 
 SEPARATE_EF = True
